[refactor]detach functions about drag and sort patterns
